LayoutTests/media/modern-media-controls/buttons-container/buttons-container-constructor.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>ButtonsContainer</code> constructor.");

const defaultContainer = new ButtonsContainer;
shouldBeEqualToString("defaultContainer.element.localName", "div");
shouldBeEqualToString("defaultContainer.element.className.trim()", "buttons-container");
shouldBe("defaultContainer.margin", "0");
shouldBe("defaultContainer.padding", "0");
shouldBe("defaultContainer.buttons", "[]");

const buttons = [new Button, new Button];
const containerWithParameters = new ButtonsContainer({
    margin: 10,
    padding: 20,
    buttons: buttons,
    cssClassName: "foo"
});
shouldBeEqualToString("containerWithParameters.element.localName", "div");
shouldBeTrue("containerWithParameters.element.classList.contains('buttons-container')");
shouldBeTrue("containerWithParameters.element.classList.contains('foo')");
shouldBe("containerWithParameters.margin", "10");
shouldBe("containerWithParameters.padding", "20");
shouldBe("containerWithParameters.buttons", "buttons");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
